Feat : 판매자 랭킹 페이지 조회 - Cursor 페이징 방식 사용

+++ src/main/java/com/promptoven/sellerbatchservice/application/batch/BatchService.java
@@ -118,7 +118,7 @@ public class BatchService {
 
         int rank = 1;
         for (AggregateEntity aggregate : allAggregates) {
-            aggregate.updateRank(rank++);
+            aggregate.updateRank((long) rank++);
         }
 
         aggregateRepository.saveAll(allAggregates);
